+++ src/main/java/Application.java
@@ -1,7 +1,7 @@
-import chess.controller.Controller;
+import chess.controller.console.ConsoleController;
 
 public class Application {
     public static void main(String[] args) {
-        Controller.run();
+        ConsoleController.run();
     }
 }
