+++ src/main/java/chess/controller/console/ConsoleController.java
@@ -1,20 +1,20 @@
-package chess.controller;
+package chess.controller.console;
 
-import chess.controller.command.Command;
-import chess.controller.command.CommandRouter;
+import chess.controller.console.command.Command;
+import chess.controller.console.command.CommandRouter;
 import chess.domain.manager.ChessGameManager;
 import chess.domain.manager.ChessGameManagerFactory;
 import chess.util.Repeater;
 import chess.view.InputView;
 import chess.view.OutputView;
 
-public class Controller {
+public class ConsoleController {
     public static void run() {
         OutputView.printInitialMessage();
         ChessGameManager chessGameManager = ChessGameManagerFactory.createNotStartedGameManager();
-        Command command;
+
         do {
-            command = Repeater.repeatOnError(() -> CommandRouter.findByInputCommand(InputView.getCommand()));
+            Command command = Repeater.repeatOnError(() -> CommandRouter.findByInputCommand(InputView.getCommand()));
             chessGameManager = executeCommandOrPassOnError(chessGameManager, command);
         } while (chessGameManager.isNotEnd());
