+++ src/main/java/chess/controller/console/command/CommandRouter.java
@@ -1,4 +1,4 @@
-package chess.controller.command;
+package chess.controller.console.command;
 
 import java.util.Arrays;
 import java.util.List;
