+++ src/main/java/chess/controller/console/command/Status.java
@@ -1,4 +1,4 @@
-package chess.controller.command;
+package chess.controller.console.command;
 
 import chess.domain.manager.ChessGameManager;
 import chess.domain.statistics.ChessGameStatistics;
