+++ src/test/java/chess/controller/console/command/CommandRouterTest.java
@@ -1,4 +1,4 @@
-package chess.controller.command;
+package chess.controller.console.command;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
